+++ MVVMArchitectureTemplateTests/Utility/NSObject+ExtensionTest.swift
@@ -1,0 +1,8 @@
+@testable import MVVMArchitectureTemplate
+import XCTest
+
+final class NSObjectExtensionTest: XCTestCase {
+    func test_className_クラス名がStringで取得できること() {
+        XCTAssertEqual(NSObjectExtensionTest.className, "NSObjectExtensionTest")
+    }
+}
